feat(mkmr): parse configuration to get personal access token
This adds configuration parsing by using the find_config, then passing
it to the call to gitlab.Gitlab.from_config so we can do API calls to
create the merge request.

The section relevant to mkmr is derived from the host of the upstream
remote url, so the your upstream remote url is https://gitlab.alpinelinux.org
then your section will be gitlab.alpinelinux.org.

mkmr also does some validation before passing the configuration to the
gitlab.Gitlab.from_config call, this is done so we can provide vital
information in a friendly manner.

The first check is if the section exists, it will exit with a long
message asking the user to write out the section name and the
private_token.

The second check is if the url option exists in the section mkmr is
looking at. If the url option does not exist then mkmr will derive it
from the host of the upstream remote and write it to the config, then
write it to the configuration file.

The third check is if the private_token option exists, if it does not
exist then it will warn the user with a long message and ask them to
write out the private_token.