feat(mkmr): make find_config() accept path as a positional argument
This makes the find_config option accept a positional argument which
represent the path to search for a configuration file.

If the positional argument is None then search in the normal locations
as the function does before this commit. If the positional argument is
not None then check if the path is a file and return it if it is, if it
is not a file then raise a ValueError pointing to the path